Fix reference renames
