+++ piezo_web_app/PiezoWebApp/tests/services/spark_application_builder/manifest_populator.py
@@ -2,7 +2,7 @@ import pytest
 import unittest
 
 
-from PiezoWebApp.src.services.application_builder.manifest_populator import ManifestPopulator
+from PiezoWebApp.src.services.spark_application_builder.manifest_populator import ManifestPopulator
 
 
 class TestTemplatePopulator(unittest.TestCase):
